+++ src/test/java/com/woowacamp/soolsool/core/receipt/repository/ReceiptStatusRepositoryTest.java
@@ -9,15 +9,17 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
 
 @DataJpaTest
 @Sql("/receipt-type.sql")
 @DisplayName("통합 테스트: ReceiptStatusRepository")
+@Import(ReceiptStatusCache.class)
 class ReceiptStatusRepositoryTest {
 
     @Autowired
-    ReceiptStatusRepository receiptStatusRepository;
+    ReceiptStatusCache receiptStatusRepository;
 
     @Test
     @DisplayName("주문서 상태를 주문서 상태 타입으로 검색한다.")
